Flex Box and Nav Bootstrap
Learned about flex box and wrapping  and how to move elements to different locations. we also learned how to align navs, and learned how to make elements responsive to different sizes of windows. We also worked with rows and column sizes to determine how much space elements will take on the webpage.